fix: drive the simulated servo from TIM2_CH2 on PA1
Proteus's STM32 model does not implement TIM4_CH3 on PB8, so the servo
never moved in simulation. Both working reference projects drive it from
TIM2_CH2 / PA1, and TIM2 is implemented -- the 10 ms system tick has
always run on it. Pulse widths stay 500 / 2500 us on both sides.

The simulation and hardware builds now use different pins. The hardware
build is unchanged in behaviour; every difference is confined to
`#if USE_SOFT_I2C` in bsp/, plus the ADC channel table in gas.c.

                servo   timer      MQ6   valve LED
  soft (Proteus)  PA1   TIM2_CH2   PA5   PB9
  hw (board)      PB8   TIM4_CH3   PA1   PA5

PA1 had to be freed for the servo, and the 48-pin package has no spare
analog pin -- all ten ADC-capable pins are occupied -- so MQ6 took PA5
and the valve LED moved to PB9. bsp_led_init() is new because the valve
LED is no longer configured by CubeMX in the simulation build.

TIM2's period becomes 20 ms to give the servo 50 Hz, so the tick ISR
calls app_tick_isr() twice to keep the 10 ms cadence. Everything else in
app.c times off HAL_GetTick(), which is unaffected.

Also in this batch:

- Buzzer setting re-encoded as -1=ALWAYS / 0=OFF / 1..60 seconds and
  cycled as a ring. The old encoding put ALWAYS next to OFF on the
  number line, so stepping it gave a counter-intuitive order. Config
  format is now v8; v6 and v7 are still read and upgraded in place.
- Recovery thresholds pinned to WARNING 80% / SAFE 70% / RELEASE 75%,
  matching what the hardware docs always claimed.
- Display pushes the whole screen when the page or the browsed history
  entry changes; it used to take eight 88 ms frames and tear on the way.
  The history entry is now cached instead of being re-read from EEPROM
  on every 200 ms redraw.
- Soft I2C calls HAL_I2C_DeInit() before taking the pins over.
- UART error callback writes the HAL_UART_Receive_IT() result back into
  `ready` rather than discarding it, so a failed restart no longer looks
  like a healthy port.
- build_keil.ps1 no longer treats UV4 warnings (exit 1) as failure. It
  used to skip artifact collection, leaving a stale HEX in Artifacts that
  Proteus would silently keep running.
- serial_check.py expectations updated for the new buzzer encoding.
- Docs updated for the new pinout; the Proteus README now records why the
  two builds differ.

Verified: host tests 7/7 on both EEPROM models, four GCC presets and both
Keil targets build with zero warnings, and each `#if` branch was checked
at object-file level (htim2 vs htim4, channel table bytes, bsp_led_init
disassembly) so a missing define cannot silently collapse the two builds
into one.

+++ stm32f103/app/gas/gas.h
@@ -69,7 +69,9 @@ void gas_key(gas_t *m, unsigned key, uint32_t now);
 bool gas_set_threshold(gas_t *m, gas_channel_t channel, uint16_t value, uint32_t now);
 bool gas_set_period(gas_t *m, uint16_t ms, uint32_t now);
 
-bool gas_set_buzzer(gas_t *m, uint16_t value, uint32_t now);
+/* value 取 int 而非 int8_t：调用者递进来的越界值要先被看见，不能被收窄成
+ * 某个合法档位。合法取值见 config.h：-1=ALWAYS, 0=OFF, 1..60=秒。 */
+bool gas_set_buzzer(gas_t *m, int value, uint32_t now);
 
 void gas_close_valve(gas_t *m, uint32_t now);
 bool gas_valve_open(const gas_t *m);
